requirements.txt
Django
django-autoslug
unidecode
django-hashedfilenamestorage
better_profanity
channels
channels_redis
